Add submodules support
Regular git-submodule cannot be used in gear since submodule objects are
not in the object database. From the current object database, there is
only a reference to a hash in another repository. Therefore, when
cloning such a repository, not all necessary objects will be available.

This utility allows you to save and restore the necessary objects as a
bundle. The utility saves all submodules recursively.

Signed-off-by: Alexey Gladkov <legion@altlinux.org>