gear-store-submodules: new utility to maintain submodules
Regular git-submodule cannot be used in gear since submodule objects are
not stored in the object database, there is just a reference to a hash
in another repository.  Therefore, cloning such a repository does not
guarantee that all necessary objects are available in the clone.

Introduce gear-store-submodules to save and restore the necessary
objects as bundles.  This utility saves all submodules recursively.

Signed-off-by: Alexey Gladkov <legion@altlinux.org>
Co-authored-by: Dmitry V. Levin <ldv@altlinux.org>